Move db modal to discover/database folder
