+++ packages/teleport/src/Discover/Database/AddDatabaseModal/AddDatabase.story.test.tsx
